Animated PNG screen recorder: capture mouse cursor too
Windows doesn't actually allow you to capture the mouse cursor as part of screenshots, so we use the standard workaround of querying system cursor position and manually rendering the cursor where the system says it should be.  This doesn't *perfectly* mimic the actual system cursor (especially if the user is running a weird custom theme on an old OS like XP), but for normal system cursors, it'll replicate everything but the drop-shadow with pixel-accuracy.