APNG screen recorder: new wizard-esque UI
Animated screen recording is now a 2-step process.

Step 1: choose all recording settings.  (There aren't many now, but there will be more soon!)

Upon hitting "OK", the user is prompted for a destination filename.  I proptotyped a lot of different ways to handle this, but I believe handling the destination filename upon hitting OK is most intuitive, because it lets me provide a standard common-dialog box with built-in "do you want to overwrite this file?" detection, just like all other dialogs in the program.

Step 2: if everything from Step 1 validates (and the user successfully selects a destination file from the common dialog), the settings window is hidden and the actual screen recording window is raised.  This dialog no longer provides any recording options - instead, it's just a capture area with a "Start" and "Cancel" button.  This is much simpler to implement (especially when the window gets resized), and it allows the screen recording to start immediately, no delays, when the user hits "Start recording".  This is also more intuitive than LICECap, which prompts the user for a filename upon hitting "start" - a behavior I've never fully gotten used to.

This UI approach is also amenable to full-screen captures.  I will shortly add an option for that to the settings window, and when selected, it will prevent the 2nd "window region" dialog from raising at all.  Instead, the full-screen recording will just start right away.