APNG screen recorder: lots of fixes, improvements
This is slowly stumbling toward a functional tool!

Last window position is now remembered between sessions.

PD's main window is auto-hidden prior to launching the capture window.

Also, SetWindowRgn is now used (instead of SetLayeredWindowAttributes) because Win 10 gave me endless grief with the old system.  When the main app window is minimized, Windows mis-reports mouse coordinates by the height of a window titlebar on child windows, *if* layered window transparency is used.  I tried for several days to find a way around this and could not, but it's a dealbreaker as it prevents the user from clicking the buttons beneath the capture area.  (And if I move the buttons elsewhere, it prevents the user from resizing the window via its bottom edge unless I leave a titlebar-sized margin down there.)

It's (unlikely but) possible MS will fix this in a future Win 10 release, so I've left the original code with a toggle to reinstate it if it becomes workable in the future.  In the meantime, however, I'm falling back to the legacy SetWindowRgn method because it gives me no grief on Win 10, despite performing an identical function.