Fix Shift key behavior when resizing layers
Fixes #306.  Many thanks to @Fivavoa for catching and reporting!

The aspect ratio calculation was wrongly calculating layer HEIGHT instead of layer BOTTOM.  This makes no difference when a layer is positioned at (or near) [0, 0], but when it's near the bottom-right of the canvas, the layer gets crushed down to tiny sizes!

(facepalm) so sorry for not catching this earlier