Lots of code cleanup
This cleanup is focused on removal of dead code (including unused APIs) across the project.  3rd-party code has been the biggest target as some wrappers (like FreeImage) expose a ton of public types and functions, most of which are unused in PD.

I've also removed a 10-year-old dialog originally designed as part of a File > Print overhaul, but which I've repeatedly put off in the adjoining years.  A tool like that is a massive project and I question the value of even pursuing a project like that anymore, given that the built-in Windows print wrapper seems to work okay for most needs.

idk.  "Maybe someday" remains my rallying cry... ;)