Check main window minimization state before raising dialogs
Fixes #307.  Thank you to @Fivavoa for catching and reporting!

Quick explanation: PhotoDemon uses a central manager for raising all dialogs.  This manager is responsible for applying visual styles to the window and determining its initial position (usually centered over the main PD window).

I thought it impossible for PD to raise a dialog while being minimized, but as @Fivavoa reports, the "this image has unsaved changes" dialog can be raised from a minimized state by right-clicking PD on the taskbar and attempting to close it, while an image with unsaved changes is active in the program.

This edge-case creates a weird bug where the dialog manager attempts to center the dialog over the main window (whose size and position is calculated as all-zeroes), resulting in a 0x0 dialog being created at position [0, 0].  The main window locks as it goes into a modal dialog loop, and because the newly raised dialog is inaccessible, the program is effectively frozen.  D'oh!

Easy, universal fix is to simply check main window state before raising *any* dialog, and forcibly restore it to normal state before entering a modal loop.