Move/size tool: remember some settings between sessions
Relates to #309.  Thank you to @mizka76 for catching and reporting!

Nearly all tools in PhotoDemon remember their last-used settings.  The Move/Size tool is a rare exception, as the majority of the tool's UI is tied to properties of the currently active layer (e.g. layer size, position, etc).

That said, there is a single panel of settings that relate only to the *tool's* behavior (e.g. "show layer boundaries") - and those settings *should* be saved between sessions.

To accomplish this, I've added some new toggles to the last-used settings manager.  Dialogs can now forcibly suspend automatic settings storage, and instead rely solely on manual load/save of custom settings (an arrangement required for this tool to work correctly, so it doesn't automatically load settings for the last-accessed layer from the previous session!)