Fix auto-crop mistakenly triggering on some Image > Crop code paths
Fixes #312.  Thank you to @jpbro for catching and reporting this troublesome bug!

Before this change, PhotoDemon's crop worked as follows.  For each layer in the crop, we would...

1) Pad the layer to the size of the image
2) Mask all pixels outside the current selection (including "partially" selected pixels via transparency)
3) Auto-crop the layer by trimming all fully transparent borders from (2)

When working with 24-bpp layers, this worked brilliantly!  The code was also quite compact as I could use existing functions to do all the messy bits.

Unfortunately, when working with 32-bpp layers, step (3) fails miserably because the original layer may have transparent regions that should be retained post-crop.  We can't just auto-trim them.

A new, correctly working approach is more complex.  Step (3) above is now replaced with...

3) Calculate an intersecting region of the pre-crop layer and the selection region
4) Create a new, blank layer at the intersection size calculated by (3)
5) Copy the post-crop layer into the new, blank layer created in (4), with offsets calculated based on relationships between the pre-crop layer's position, the selection's position, and the intersecting position post-crop.  (These vary based on whole-image vs single-layer crops.)

This produces a correct result on 32-bpp layers, and quick testing leaves me (mostly?) confident that it works across all selection types, including complex regions.

Unfortunately, some special cases needed to be worked around in this new arrangement - for example, when cropping a multi-layer image, what should PD do with layers that lie completely outside the crop region?  The old code handled this automatically (but not elegantly) - when the auto-cropper detected a source layer that was fully transparent, it would simply blank out the layer and leave it at its temporary, padded-to-full-image-size state.  Not ideal.

Now, I've fixed it so that layers lying completely outside the crop area are instead blanked out and resized to the match the newly cropped image size.  This seems more sensible, and it has the bonus side-effect of reducing memory requirements relative to the old method.  (Other software - e.g. GIMP - instead deletes those layers entirely... which I have mixed feelings about.  I can copy that behavior if desirable, but for now I'd rather err on the side of "leave the layers in the image, and just make 'em transparent")

Note also that PD has an entirely separate pathway for "non-destructive" crops, and that code wasn't touched here.  That pathway is triggered by crop operations using a rectangular selection with no antialiasing or feathering or rounded corners, on an image with no non-destructive transforms active.  In that special case, PD can just modify layer offsets to implement the crop, without actually modifying any pixel data.  That pathway will still be used when all those criteria are met.