Animated images: navigator slider now activates frame upon mouse release
This behavior almost certainly needs to be exposed via property, as it has repercussions for animation work (where it's common to "blip" back and forth between two frames to see how they look in animation, and activating the "blipped" layer could be an unwanted side-effect).
 But seeing as I don't have animation settings anywhere at present, for 8.0's release I'm comfortable making this the default behavior.

Note that mouse-dragging the slider only *activates* the selected layer, e.g. it's the exact same as clicking that layer in the layer box.  It does *not* make the layer visible (if it's invisible) as that would add a change to the Undo/Redo stack, which makes the operation non-destructive and has a bunch of other repercussions.

As part of this commit, I've also tweaked the way the animation control tracks frame time.  This is necessary for a suggestion in #318, "showing the total animation time" somewhere... a problem that I've yet to devise a good solution to, at least before 8.0 releases (all of my preferred solutions require translation work, alas).  The navigation control can now reliably tally total animation time, at least.

This change relates to #318.  Thank you to @jpbro for suggesting it!