APNG screen recorder: after recording, add the file to the File > Open Recent menu
Recorded APNGs are now automatically added to the top of the Open > Recent files submenu.  (Said another way, after recording you can use the Ctrl+0 shortcut to immediately load your recording.)

Relates to #318.  Thank you to @jpbro for pointing me in this direction!

(Post-8.0, auto-opening the recording in PD should be added as a preference to the pre-recording options dialog, where it makes more sense.  That said, I still like this behavior as it will always provide a fallback way to quickly retrieve your latest recording.)