Lasso selections: highlight the initial point w/ the current accent color
I went back and forth on whether to do this (for both lasso and polygon points), as it's a bit of inconsistent UI.  Filling a UI element with the current accent color is usually reserved for MouseOver and focus communication, which is this is not - it's just to orient the user as to where their current selection began, so they know how to terminate their in-progress selection.

But I don't have a better way to communicate that at present (an animation would probably be ideal, but I'm not implementing that before 8.0 release), so we'll go with this.  It's worked well enough for polygons for a few years, so we may as well mimic the same thing with lasso selections for now.

Thank you to @dusko23 for the suggestion.