Add mousewheel support to spin controls
Because of the way mouse events propagate between child and parent windows, mousewheel-ing over the edit box, spin controls, or reset button on a spin control will trigger a change event.

Some hasty research leads me to believe this is likely okay across all locales, but if for some reason this screws with an IME or other edit box input method, please let me know so I can revert it!

Thank you to @dusko23 for the suggestion.