Edit > Paste as layer: if a selection is active, align the pasted layer...
...with the selection.  This means that making a selection, then hitting copy+paste, will position the new layer perfectly over the existing selection.

This appears to be how most other editors behave, and I apologize for not noticing this and implementing it sooner!

(Note for 8.0 release: the interaction between this and the recently added "paste at cursor position if cursor is over image" behavior is complex and unintuitive.  I should probably look at making "paste to cursor" a dedicated Edit menu option, a la Krita)