Clipboard operations: allow them to be recorded during macros
It's hard to explain how this solves #320, but it does!

Basically I had concerns with allowing certain operations to be recorded during macros, specifically Edit > Paste in this context.  Why?
 If the clipboard's contents are not valid when that macro is played-back, prompts will be raised during macro playback which can ruin the batch processing experience.

That said, some elements of PD's central command processor rely on macro state to know how to handle things like Undo/Redo data (e.g. Undo/Redo generation is suspended during batch processing for performance reasons).  This interplay is complicated and I don't want to mess with it so close to 8.0's release, so an easier fix is to simply reinstate Edit > Copy/Paste commands during macro recording, which is probably preferable anyway as those ops may be useful to automate for some workflows?

Apologies for the lengthy explanation.  Some PD interactions are weird!

Thank you to @mizka76 for catching and reporting.  He has been an all-star tester on this release!