Color picker: new option for "switch to previous tool" after clicking
Relates to #316.  Thank you to @jpbro for the great suggestion!

I'd still like to mimic Photoshop's support for holding the ALT key while using any paint tool to temporarily switch to the color picker (then switch back when ALT is released).  However, this requires some changes to the way PD loads/unloads toolboxes, so it'll be slightly more involved to implement.

(and oh yeah - thus commences work on v8.2!  No rest for the wicked... ;)