New toolpanel caching strategy
PhotoDemon has a lot of (complicated) option panels for its various on-canvas tools.  To improve startup performance, tool option panels are not loaded until the user actually requests that tool.  This results in a slight delay when activating a tool, as the associated options panel is loaded, restored to its last-used settings, themed, positioned, and ultimately rendered to the screen.

In 8.0, when switching to a new tool, PD completely unloads the previous tool's option panel.  This keeps resource usage extremely low, at some cost to performance (as there's a delay if the user returns to that old tool, because everything has to be re-loaded from scratch).

This approach causes problems in implementing a better version of #316, specifically the ability to hold down a key like ALT when painting to temporarily switch to the color picker.  (Then you'd return to the original paint tool when ALT is released.)  The color picker tool loads pretty quickly, but there can still be a subtle delay, and it's off-putting.  To make this feature work well, the color picker needs to load instantaneously.

So I'm trying a new strategy.  As of this commit, PD still doesn't load any toolpanels when the program starts.  BUT - once a tool panel has been loaded during a session, PD leaves it in-memory, so that it can load it instantly if the user returns to it.

This seems like a small change, but there are potential issues with the way option panels are managed since they use a lot of low-level APIs to modify their behavior when they're first displayed.  (Freeing them completely and reloading them completely on each access bypassed some of these issues.)  It will also cause an increase in resource usage if multiple tools are used in a session, but compared to any other photo editor, this impact should still be pretty minimal.

I think I've restructured the code in a way that prevents any problems, but more testing is obviously required!