Paint tools: ALT key now (temporarily) activates Color Picker tool
...and when ALT is released, you'll return to whatever paint tool you were using.

This mirrors Photoshop's behavior, right down to the ugly complexities required to prevent the ALT key from switching focus to the menu bar.  I've tested (somewhat thoroughly?) to make sure this doesn't mess with hotkeys that use the ALT key, so all should be well.

Relates to #316.  Thank you to @jpbro for the suggestion!  This is an awesome usability improvement.