Animated images: many engine improvements, new Image > Animation menu
PhotoDemon was never intended to be an animated file editor, but now that animated GIF and PNG files have found a home here, a lot of features need to be expanded to make animations "first-class" citizens.

Unfortunately, 8.0's release got delayed so many times that I eventually had to shove it out the door without fully integrating these features into the project.  Time to start rectifying that!

In this commit, I'm landing a lot of new animation-related tech.  Animated files are much more strenuous on resource management than normal images, because they can have dozens (even hundreds) of frames, all at the full size of the image.  Managing these is a different beast, and I'm working on improving PD's behavior.

To that end, there's a new Image > Animation menu that exposes a (only partially working) dialog for animation features.  What's relevant at the moment is that this dialog uses a new caching strategy for generating spritesheets on-the-fly for its animation previews.  PD already used dynamic spritesheets, but the caching strategy was geared much more towards small collections of small images, i.e. UI elements like button icons.  Something like an animated screen capture is a different beast, because each frame can be closer to photo sizes, so spritesheets are not the resource-saver they'd otherwise be.

For this animation dialog, the new spritesheet generator aims for specific memory-limits per sheet rather than a fixed number of sprites per sheet.  And instead of building an ever-larger sheet, it instead builds discrete subsheets if the main sheet grows too large.  This is a great strategy for animations, and it makes a meaningful dent in resource usage when frame count is very high.

PD also caches frame times directly inside layer objects now.  Previously, it would attempt to parse this from layer names; this was a cheap and dirty strategy because I was lazy and just wanted something that "worked".  The new strategy is much more elegant (and faster!) but it required new synchronization code to detect changes to layer names when frame times are involved, since we need to add Undo/Redo flags for frame time changes.  Hence the large amount of files touched in this commit; the import engine in particular needs to set the new frame time flags correctly.

The new Image > Animation dialog is somewhat unique in that it's resizable.  PD's windows are mostly locked to set sizes (to make my life easier), but I hope to remedy this in the future.  A new framework is needed for remembering window sizes and positioning windows intelligently (inc. accounting for things like detached displays since the last session), but I'm aware of all that and will be tackling it in time.  Anyway, this is relevant here because the animation window will automatically scale the animation as the window is resized - nice for large animations in particular, although I'd still probably recommend using a web browser for idealized previews (since that's where most animated images ultimately live).

Relating to great input from @jpbro in #318, the navigation panel on the main window now has an expanded tooltip for the slider control.  Hover it to see current animation frame (and total animation frames), as well as the current frame's start time (and total animation time).  I don't know if this is the best place to put this information... feedback welcome!

Not everything in the new Image > Animation window is working yet.  In particular, animation speed controls are not yet synced to the parent image upon hitting OK.  That's coming soon.