Image > Animation menu: wrap up work (for now)
This tool can now be used to switch images between static and animated modes, with some support for modifying things like frame rate across the image.

This feature was a blocker for my next task: silently activating the Export > Animated GIF/PNG menu items when the user calls Save/Save As with an animated image.  This wasn't done for 8.0's release because PhotoDemon couldn't reliably know whether an image was animated or not, so Save commands always mapped to static image save formats.  Instead, the user was forced to use Export > Animated commands to save animated images, which is a big usability fail.

Now, however, all the internal flags are in place to make switching between static/animated images when saving automatic.