Ctrl+ OR Alt+ OR Shift+Click layer visibility icon to show *only* that layer
- Photoshop uses Alt+Click
- GIMP users Shift+Click
- Who knows what other software uses?

PhotoDemon doesn't care - when any modifier is pressed when clicking the "visibility" (eyeball) icon on the layer box, PD interprets it as "make this the ONLY visible layer in the image".

If that layer is already the only visible layer in the image, PD will instead "show all layers".  This mirrors identical behavior in Photoshop and GIMP.

This feature was suggested by @jpbro in #318.  Thank you to him for this excellent usability improvement!