Language editor: switch to DeepL.com for estimating missing translations
Some years back, one of PD's volunteer translators suggested that PD's built-in language editor should estimate missing translations using an online service (originally Google Translate).  If the translation estimate was terrible, no loss, but if it was okay, it could save translators time because they could simply tap "enter" and move onto the next one.  (Or they could make minor tweaks rather than typing out a whole translation from scratch.)

The problem with this approach is that I'm not using the official Google Translate API, so any changes to their page layout breaks this "feature".  (I genuinely hesitate to call it a "feature" because it's so brittle, lol.)

Anyway, a better solution would be to use official APIs, but no libraries exist for this in VB6 and I'm too lazy to write my own as I'm not even sure how useful it is.  (Based on my limited sampling, most of PD's translators just use a text editor, not PD's built-in tool.)

Anyway, that's the history here.  As of this commit, I've switched PD's (currently non-functional) Google Translate "estimations" over to DeepL.com, based on the advice of one of my translators.  DeepL doesn't support as many target languages as Google, but I hear the ones it *does* support work significantly better - so a win for everyone, I hope!