Tools > Preferences: new preference for canvas backcolor
For decades, Photoshop didn't provide a preference for this, but they finally relented and added it (I believe?) when they also added built-in support for dark UI themes.  I've received a request to do the same in PD, and I'm happy to oblige!  (Previously, I tied canvas backcolor to the current UI theme, but there is wisdom in separating it out.)

PD still defaults to neutral gray, but you can now manually override this with any color you want in the Options > Interface panel.

While I was here, I also started reworking the Options panel to better reflect UI patterns elsewhere in PD.  I want to make the Options panel more visual and less word-y, since this dialog comprises a disproportionate burden on PD's volunteer translators.  Expect more changes in the coming weeks.