Fix critical crash when exporting monochrome DIBs
Relates to #327.  Many thanks to @KeratosK for catching this.

I'm deeply frustrated that 8.2 shipped with this bug.  It was a silly oversight; basically, PD's PNG exporter is very aggressive about optimizing exported PNGs.  With grayscale images (including monochrome ones), the PNG exporter will flatten the image before handing it off to the grayscale analyzer IF it doesn't detect any meaningful alpha values in the image.   Unfortunately, the grayscale analyzer was assuming that it would always be handed un-flattened data, and the mismatch was causing crashes.

Dammit, there's always something.  I brought this on myself by aggressively testing PNG images with varying alpha, since that's the use-case I really wanted to optimize since no other software does it well... and in doing so, I accidentally broke basic monochrome images.  AAARRRGGGHHH

Anyway, the problem is now fixed.  Guess we're gonna need an 8.4 release in the coming weeks *sigh*