Finalize work on new "quick start" panel
I like how this turned out!  PhotoDemon starts up nice and quick, but if it forces the user to dive into menus to load/create images, its fast start-up speed advantage isn't as meaningful.

I'm also not the biggest fan of apps that start you into a blank document of some arbitrary size.  I *never* need an image that size, so generating a "default" image is a waste of time and resources.

PD tries to be better about this by not allocating any special memory for a "dummy" start-up image.  Instead, it gives quick links to "new image" settings (including opening and importing images), as well as an optional list of recently accessed images.  For all of these controls, it reuses existing assets from related menus so there is no penalty to start-up speed.

The layout is a little non-traditional at present.  I'm gonna let it simmer for a bit before making a final decision on how it should work.  (I may top-align both columns to the same position, instead of allowing them to float independently as they do now... idk, we'll see.)