Fix crash in batch processor when setting export options for PNG and GIF files
Relates to #330 .  Many thanks to @Fivavoa for reporting and @mizka76 for confirming!

The bug was caused by new logic that automatically switches between animated and non-animated export screens, based on the contents of the current image.  When setting export settings at the start of a batch conversion, there is no active image - so the logic was crashing when it attempted to access settings for the currently active image.