Get rest of plumbing in place for async multi-session communication
Relates to #328

It works!  If PD is already running, subsequent sessions can now forward their command-lines to the first session, and it will auto-load any+all images they were passed.

I'm not actually activating this yet because I am first going to add a user pref for it; once that's available, I'll toggle the behavior on, with a few added safety fixes so that bad things can't happen (like loading new images in the middle of a running batch conversion).