Tools > Options > Loading: New single-instance app preference
Relates to #328.  Thank you to @g-busch for the original suggestion!

I debated where to stick this new preference, but ultimately it really only applies to image loading (since that's when the app decides whether to reuse an existing instance or start a new one), so Tools > Options > Loading it is.

By default, PD is set for multi-instance mode, same as it's always been.  What this means is if you drag an image onto PhotoDemon.exe, it will start a new app instance with that image automatically loaded, even if a copy of PhotoDemon is already running.

If you change the toggle to "single-instance", dragging images onto a PhotoDemon.exe instance (or loading them any other way that triggers PD to start) will instead reuse a single PhotoDemon instance for all loaded images.  This is how Photoshop, Microsoft Office, and any number of other programs work, so it's great to now have that option available for PD.

Note that some situations prevent single-instance mode from triggering, by design.  For example, if you are in the midst of a batch process (or anything else involving a live modal dialog), PhotoDemon will launch a new instance regardless of your preference; this avoids the possibility of screwing up actively running operations in your first instance, while also providing immediate access to any loaded images.

As for technical details: I'm doing interop between app instances using named pipes, which is almost certainly overkill for this, but it was a great opportunity to write new classes for something I haven't done in VB6 before.  I've been meaning to test out named pipes more in PD to see if they might work as an IPC mechanism for future performance improvements, and so far, the results look promising.  (That said, the current pipe code uses polling instead of a full overlapped implementation - one thing at a time!)