pdButtonToolbox: add Shift parameter to Click() events, so I can support...
...Shift/Ctrl/Alt clicks for alternate behavior.

GIMP uses this to excellent effect for the buttons at the bottom of its layer toolbox.  For example, you can click the "up" button to move a layer up, or Shift+Click it to move a layer to the top of the stack.  Similarly, click "new layer" to bring up a new layer dialog, or Shift+Click to bypass the dialog and auto-create a new layer with your last-used settings.  Very clever, and descriptive tooltips explain these alternate behaviors in detail.

Given how sacred space is in PD's right toolbox, I'm eager to steal this idea.

This UI element is used everywhere, so event declarations had to be changed to match - hence the large number of touched files.

Thank you also to Roy K for an updated German localization!