New Effects > Edge > Gradient flow tool
I'm doing some promising prototyping on a magnetic lasso tool, but testing the results interactively has proven difficult because it's hard to see how PD si analyzing object boundaries when attempting to "magnetize" the lasso to nearby boundaries.

So I wrote a visualizer for showing gradient flow calculations in the current image, and it's been so helpful that I may as well expose it to users.  I've cleaned up the prototype code a bit and added some options to make output easier to understand.

In this context, "gradient" doesn't refer to the Photoshop-esque term of smoothly fading between colors - instead it refers to something closer to edge-detection:

https://en.wikipedia.org/wiki/Image_gradient

PD uses various internal functions to calculate gradient magnitude and direction for each pixel in an image, and those results are used as inputs for all kinds of other effects.  The purpose of this new tool is to paint those gradient calculations directly onto the image, so you can see where object boundaries lay, at least as far as PD knows.

The most interesting setting on this new tool is the "direction" output, which paints little arrows across the image, with each arrow's direction and length showing where PD detects an object boundary (and how strong it thinks that boundary is).  For an obvious example of how it works, paint a black-and-white reflection gradient using PD's gradient tool, then run this new Effects > Edge > Gradient tool on the result, with the "direction" setting active.  (If your base image is large, you may need to crank up the boost and/or sample radius settings.)