Tools > Options: rework Saving preferences
Relates to #335.  Thank you to @risk00 for motivating me to change this sooner rather than later!

PhotoDemon now defaults to saving files to the current file's folder (if it exists).  Pro users can still toggle to my preferred behavior, which is suggesting the last-used folder instead.

While I was here, I rewrote the text for several other preferences in this screen (my apologies, translators).  This panel was extremely wordy, and I'm trying to simplify and unify text wherever I can.  (I'm also trying to reduce usage of "you" or "me" in UI elements, as those aren't used consistently throughout PD - so that's the reason for some other text changes.)

I think this panel is much clearer as a result, especially for beginners.