Effects > Artistic > Color Pencil: start work on a new implementation
One of the goals for the next release is to provide a more comprehensive selection of Artistic Filters, closer to what Photoshop provides.  PD has gained many internal "painting" features over the past few years, and these types of filters should now be feasible to reproduce (I think - time will tell, ha).

Anyway, this is a very preliminary implementation and nowhere near ready for primetime, but I need to upload a new nightly build so this is coming along for the ride.