Adjustments > Levels: finish fixes required for run-time resizing
This dialog should probably be rebuilt from the ground-up (it's just not designed very well), but at least it's now working again.  The whole dialog is running off PD's UI toolkit so it should perform better than before as well.

Next up, fixing things in the Adjustments > Curves dialog that broke because of changes here.