Adjustments > Curves: big ol' UI overhaul
Relates to #261

Recent fixes to the resizability (probably not a word?) of the Levels dialog broke a bunch of UI behavior in the Curves dialog.  (These tools share a lot of under-the-hood code.)  These breaks were intentional, as the Curves dialog needed a rewrite to migrate its interactive "curve editor" UI to PD's internal rendering toolkit.

Such a rewrite is now complete, and the new Curves dialog is much faster, much prettier (including dark theme support), with a greatly improved and modernized codebase for its UI bits.  The new dialog fully supports run-time resizing, and its improved performance means it runs very smoothly even when maximized on a very large display.  A bunch of annoying UI quirks were also fixed while I was here, and I'm happy to say this dialog is in a much better place than it was at 8.0's release.

Anyway, this was the last dialog with serious reflow issues during resizing.  I still need to fix some non-UI dialog resize quirks on a few tools (some tools perform specialized caching against the preview window, and they need to be rewritten to handle the new possibility of the preview window's size changing at run-time), but overall, #261 is very close to "true" completion.  This is an achievement that not even Photoshop offers on most of its tools, which I'd say is a pretty big win for us.