Image/Layer > Straighten and Rotate tools: fix dialog resize behavior
Relates to #261

These dialogs serve dual-duty, since they appear in both the Image menu (where you can straighten or rotate all layers at once) and the Layer menu (for just the active layer).

When these tools are loaded, they make a preview-sized copy of the source image or layer and use that copy as the basis of the effect preview.  This provides a nice performance boost when the original image/layer is much larger than the preview window.

Both dialogs needed to be updated to cover the possibility of the preview size changing at run-time (if the dialog itself is resized).