Adjustments menu: remove Photography submenu
This submenu never made sense, since photography encompasses all adjustments in one way or another.

As for the two tools in that submenu:
- Photo filters have been moved to the color sub-menu.
- Red-eye removal has been removed.

A standalone dialog for red-eye removal was a bad choice.  I used a research paper by Microsoft as the basis of my implementation, but it never worked well on large modern images because the region detection requires way more nuance than it could provide.

A better solution is the Photoshop approach of a dedicated on-canvas tool where the user can just click eyeballs to fix 'em.  I'll add this to my todo list.