Adjustments > Photo filters: complete overhaul
This is one of those "if you can't beat 'em, join 'em" changes.

Photo filters are literal filters that you snap over the top of a camera lens.  They work by filtering (or in some cases, emphasizing) specific wavelengths of light, which allows for dramatic light capture that cannot be reproduced in post.

Photoshop provides a weird mimicry of these filters in its "Image > Adjustments > Photo Filter" tool.  I originally tried to outdo Photoshop by providing a much more comprehensive list of filters in PhotoDemon's version of the tool, complete with each filter's full technical name and descriptions.

Unfortunately, that made this tool an extraordinary translation burden, with 70 phrases comprising nearly 400 words of text for a single esoteric dialog.  (For perspective, that's more than 4% of PhotoDemon's total language file!)

This was insanity, and today I finally reworked this dialog to simply reproduce the same options as Photoshop.  That's the primary purpose of this tool, after all - to provide a "drop-in" replacement for Photoshop tutorials that require it.  It was silly of me to overcomplicate it in the first place, and I apologize to my translators for such an overkill implementation.

The new version of this tool is leaner, cleaner, and faster.  While here, I reworked the algorithm to operate exclusively in the CieLAB color space, which produces a higher-quality transform than Photoshop at an even faster speed than PD's old implementation.  The result is more dramatic lighting regardless of which "filter" you're applying, with especially improved results when trying to e.g. warm up a very cool photo (or cool a photo with very warm tones).