New function for "shuffling" floats to improve compression
Floating-point values compress very poorly, thanks to their floating-point nature.  (Bit distribution can vary greatly between neighboring values, even if the values themselves are similar.)

One way to reduce this penalty is to shuffle bytes so that the first byte of all floats are placed together in a row, then second bytes, third, etc.  This increases the chance of patterns and in the case of 3D LUT files (which I'm working on), the compression ratio goes from < 10% to 30-50% thanks to this very simple modification.

I've seen reports that shuffling nibbles will do even better, but that's quite messy in VB so I'm limiting it to whole bytes for now.